fuelclient/test-requirements.txt
-r requirements.txt
hacking==0.7
mock==1.0
nose==1.1.2
nose2==0.4.1
nose-timer==0.2.0
sphinx==1.2
rst2pdf==0.93
sphinxcontrib-plantuml==0.3
sphinxcontrib-blockdiag==1.3.0
sphinxcontrib-actdiag==0.6.0
sphinxcontrib-seqdiag==0.6.0
sphinxcontrib-nwdiag==0.7.0
webtest==2.0.14
ply==3.4
python-keystoneclient>=0.7
